+++ src/err.cpp
@@ -190,7 +190,7 @@ void zmq::win_error (char *buffer_, size_t buffer_size_)
     DWORD errcode = GetLastError ();
     DWORD rc = FormatMessageA (FORMAT_MESSAGE_FROM_SYSTEM |
         FORMAT_MESSAGE_IGNORE_INSERTS, NULL, errcode, MAKELANGID(LANG_NEUTRAL,
-        SUBLANG_DEFAULT), buffer_, buffer_size_, NULL );
+        SUBLANG_DEFAULT), buffer_, (DWORD) buffer_size_, NULL );
     zmq_assert (rc);
 }
